// src/modules/analytics/constants/analytics.enums.ts
export enum TrendPeriod {
    SIX_MONTHS = 'SIX_MONTHS',
    ONE_YEAR = 'ONE_YEAR',
    ALL_TIME = 'ALL_TIME',
    CUSTOM = 'CUSTOM',
}

export enum ActivityTimeframe {
    WEEKLY = 'WEEKLY',
    BIWEEKLY = 'BIWEEKLY',
    MONTHLY = 'MONTHLY',
    QUARTERLY = 'QUARTERLY',
    YEARLY = 'YEARLY',
}

export enum LogStatus {
    SUCCESS = 'SUCCESS',
    FLAGGED = 'FLAGGED',
}